+++ config/initializers/frilans_finans_api.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-FrilansFinansApi.configure do |config|
+FrilansFinansAPI.configure do |config|
   config.logger = Rails.logger
   config.event_logger = FrilansFinansEventLogger.new
-  config.client_klass = FrilansFinansApi::Client
+  config.client_klass = FrilansFinansAPI::Client
   config.base_uri = AppConfig.frilans_finans_base_uri
   config.client_id = AppSecrets.frilans_finans_client_id
   config.client_secret = AppSecrets.frilans_finans_client_secret