+++ lib/seeds/dev/currency_seed.rb
@@ -3,7 +3,7 @@
 module Dev
   class CurrencySeed < BaseSeed
     def self.call
-      client = FrilansFinansApi::FixtureClient.new
+      client = FrilansFinansAPI::FixtureClient.new
       FrilansFinansImporter.currencies(client: client)
     end
   end