+++ CHANGELOG.md
@@ -1098,7 +1098,7 @@ v1.60 - 2017-05-04
   - Add proper configure class for `JsonApiHelpers`
   - Proper configuration module for `FrilansFinansApi`
   - Destroy week old frilans finans api logs sweeper task
-  - Remove FrilansFinansApi::reset_config method
+  - Remove FrilansFinansAPI::reset_config method
 
 
 v1.59 - 2017-05-02
@@ -1788,7 +1788,7 @@ v1.8 - 2017-02-19
 
 v1.7.1 - 2017-02-17
 ----------
-* _FrilansFinansApi_: Add `FrilansFinansApi::User::index` method
+* _FrilansFinansApi_: Add `FrilansFinansAPI::User::index` method
 
 
 v1.7 - 2017-02-17